+++ test/integration/physics_link.cc
@@ -16,6 +16,8 @@
 */
 #include <string.h>
 
+#include "gazebo/math/Vector3Stats.hh"
+#include "gazebo/msgs/msgs.hh"
 #include "gazebo/physics/physics.hh"
 #include "test/ServerFixture.hh"
 #include "helper_physics_generator.hh"
@@ -27,14 +29,105 @@ const double g_tolerance = 1e-4;
 class PhysicsLinkTest : public ServerFixture,
                         public testing::WithParamInterface<const char*>
 {
+  /// \brief Test GetWorldAngularMomentum.
+  /// \param[in] _physicsEngine Type of physics engine to use.
+  public: void GetWorldAngularMomentum(const std::string &_physicsEngine);
+
   /// \brief Test GetWorldEnergy* functions.
   /// \param[in] _physicsEngine Type of physics engine to use.
   public: void GetWorldEnergy(const std::string &_physicsEngine);
+
+  /// \brief Test Link::GetWorldInertia* functions.
+  /// \param[in] _physicsEngine Physics engine to use.
+  public: void GetWorldInertia(const std::string &_physicsEngine);
 
   /// \brief Test velocity setting functions.
   /// \param[in] _physicsEngine Type of physics engine to use.
   public: void SetVelocity(const std::string &_physicsEngine);
 };
+
+/////////////////////////////////////////////////
+// GetWorldAngularMomentum:
+// Spawn box and verify Link::GetWorldAngularMomentum functions
+// Make dimensions unequal and give angular velocity that causes
+// gyroscopic tumbling.
+void PhysicsLinkTest::GetWorldAngularMomentum(const std::string &_physicsEngine)
+{
+  // Load a blank world (no ground plane)
+  Load("worlds/blank.world", true, _physicsEngine);
+  auto world = physics::get_world("default");
+  ASSERT_TRUE(world != NULL);
+
+  // Verify physics engine type
+  auto physics = world->GetPhysicsEngine();
+  ASSERT_TRUE(physics != NULL);
+  EXPECT_EQ(physics->GetType(), _physicsEngine);
+
+  // disable gravity
+  physics->SetGravity(math::Vector3::Zero);
+
+  physics::ModelPtr model;
+  {
+    // Box size
+    const double dx = 0.1;
+    const double dy = 0.4;
+    const double dz = 0.9;
+    const double mass = 10.0;
+
+    msgs::Model msgModel;
+    msgModel.set_name(this->GetUniqueString("model"));
+    msgs::AddBoxLink(msgModel, mass, math::Vector3(dx, dy, dz));
+    model = this->SpawnModel(msgModel);
+  }
+  ASSERT_TRUE(model != NULL);
+
+  // inertia matrix, recompute if dimensions change
+  const double Ixx = 0.80833333;
+  const double Iyy = 0.68333333;
+  const double Izz = 0.14166667;
+  const math::Matrix3 I0(Ixx, 0.0, 0.0
+                       , 0.0, Iyy, 0.0
+                       , 0.0, 0.0, Izz);
+
+  // Since Ixx > Iyy > Izz,
+  // angular velocity with large y component
+  // will cause gyroscopic tumbling
+  const math::Vector3 w0(1e-3, 1.5e0, 1.5e-2);
+  model->SetAngularVel(w0);
+
+  // Get link and verify inertia and initial velocity
+  auto link = model->GetLink();
+  ASSERT_TRUE(link != NULL);
+  ASSERT_EQ(w0, link->GetWorldAngularVel());
+  ASSERT_EQ(I0, link->GetWorldInertiaMatrix());
+
+  // Compute initial angular momentum
+  const math::Vector3 H0(I0 * w0);
+  ASSERT_EQ(H0, link->GetWorldAngularMomentum());
+  const double H0mag = H0.GetLength();
+
+  math::Vector3Stats angularMomentumError;
+  const std::string stat("maxAbs");
+  EXPECT_TRUE(angularMomentumError.InsertStatistic(stat));
+  const int steps = 5000;
+  for (int i = 0; i < steps; ++i)
+  {
+    world->Step(1);
+    math::Vector3 H = link->GetWorldAngularMomentum();
+    angularMomentumError.InsertData((H - H0) / H0mag);
+  }
+  if (_physicsEngine == "dart")
+  {
+    gzdbg << "dart has higher error for this test (see #1487), "
+          << "so a larger tolerance is used."
+          << std::endl;
+    EXPECT_LT(angularMomentumError.Mag().Map()[stat], g_tolerance * 1e3);
+  }
+  else
+  {
+    EXPECT_LT(angularMomentumError.Mag().Map()[stat], g_tolerance * 10);
+  }
+}
 
 /////////////////////////////////////////////////
 void PhysicsLinkTest::GetWorldEnergy(const std::string &_physicsEngine)
@@ -75,6 +168,165 @@ void PhysicsLinkTest::GetWorldEnergy(const std::string &_physicsEngine)
     world->Step(stepSize);
     double energy = link->GetWorldEnergy();
     EXPECT_NEAR(energy / energy0, 1.0, g_tolerance * 10);
+  }
+}
+
+/////////////////////////////////////////////////
+// GetWorldInertia:
+// Spawn boxes and verify Link::GetWorldInertia* functions
+void PhysicsLinkTest::GetWorldInertia(const std::string &_physicsEngine)
+{
+  // Load a blank world (no ground plane)
+  Load("worlds/blank.world", true, _physicsEngine);
+  auto world = physics::get_world("default");
+  ASSERT_TRUE(world != NULL);
+
+  // Verify physics engine type
+  auto physics = world->GetPhysicsEngine();
+  ASSERT_TRUE(physics != NULL);
+  EXPECT_EQ(physics->GetType(), _physicsEngine);
+
+  // disable gravity
+  physics->SetGravity(math::Vector3::Zero);
+
+  // Box size
+  const double dx = 1.0;
+  const double dy = 4.0;
+  const double dz = 9.0;
+  const double mass = 10.0;
+  const double angle = M_PI / 3.0;
+
+  const unsigned int testCases = 4;
+  for (unsigned int i = 0; i < testCases; ++i)
+  {
+    // Use msgs::AddBoxLink
+    msgs::Model msgModel;
+    math::Pose modelPose, linkPose, inertialPose;
+
+    msgModel.set_name(this->GetUniqueString("model"));
+    msgs::AddBoxLink(msgModel, mass, math::Vector3(dx, dy, dz));
+    modelPose.pos.x = i * dz;
+    modelPose.pos.z = dz;
+
+    // i=0: rotated model pose
+    //  expect inertial pose to match model pose
+    if (i == 0)
+    {
+      modelPose.rot.SetFromEuler(0.0, 0.0, angle);
+    }
+    // i=1: rotated link pose
+    //  expect inertial pose to match link pose
+    else if (i == 1)
+    {
+      linkPose.rot.SetFromEuler(0.0, 0.0, angle);
+    }
+    // i=2: rotated inertial pose
+    //  expect inertial pose to differ from link pose
+    else if (i == 2)
+    {
+      inertialPose.rot.SetFromEuler(0.0, 0.0, angle);
+    }
+    // i=3: offset inertial pose
+    //  expect inertial pose to differ from link pose
+    else if (i == 3)
+    {
+      inertialPose.pos.Set(1, 1, 1);
+    }
+
+    {
+      auto msgLink = msgModel.mutable_link(0);
+      auto msgInertial = msgLink->mutable_inertial();
+
+      msgs::Set(msgModel.mutable_pose(), modelPose);
+      msgs::Set(msgLink->mutable_pose(), linkPose);
+      msgs::Set(msgInertial->mutable_pose(), inertialPose);
+    }
+
+    auto model = this->SpawnModel(msgModel);
+    ASSERT_TRUE(model != NULL);
+
+    auto link = model->GetLink();
+    ASSERT_TRUE(link != NULL);
+
+    EXPECT_EQ(model->GetWorldPose(), modelPose);
+    EXPECT_EQ(link->GetWorldPose(), linkPose + modelPose);
+    EXPECT_EQ(link->GetWorldInertialPose(),
+              inertialPose + linkPose + modelPose);
+
+    // i=0: rotated model pose
+    //  expect inertial pose to match model pose
+    if (i == 0)
+    {
+      EXPECT_EQ(model->GetWorldPose(),
+                link->GetWorldInertialPose());
+    }
+    // i=1: rotated link pose
+    //  expect inertial pose to match link pose
+    else if (i == 1)
+    {
+      EXPECT_EQ(link->GetWorldPose(),
+                link->GetWorldInertialPose());
+    }
+    // i=2: offset and rotated inertial pose
+    //  expect inertial pose to differ from link pose
+    else if (i == 2)
+    {
+      EXPECT_EQ(link->GetWorldPose().pos,
+                link->GetWorldInertialPose().pos);
+    }
+    // i=3: offset inertial pose
+    //  expect inertial pose to differ from link pose
+    else if (i == 3)
+    {
+      EXPECT_EQ(link->GetWorldPose().pos + inertialPose.pos,
+                link->GetWorldInertialPose().pos);
+    }
+
+    // Expect rotated inertia matrix
+    math::Matrix3 inertia = link->GetWorldInertiaMatrix();
+    if (i == 3)
+    {
+      EXPECT_NEAR(inertia[0][0], 80.8333, 1e-4);
+      EXPECT_NEAR(inertia[1][1], 68.3333, 1e-4);
+      EXPECT_NEAR(inertia[2][2], 14.1667, 1e-4);
+      for (int row = 0; row < 3; ++row)
+        for (int col = 0; col < 3; ++col)
+          if (row != col)
+            EXPECT_NEAR(inertia[row][col], 0.0, g_tolerance);
+    }
+    else
+    {
+      EXPECT_NEAR(inertia[0][0], 71.4583, 1e-4);
+      EXPECT_NEAR(inertia[1][1], 77.7083, 1e-4);
+      EXPECT_NEAR(inertia[2][2], 14.1667, 1e-4);
+      EXPECT_NEAR(inertia[0][1],  5.4126, 1e-4);
+      EXPECT_NEAR(inertia[1][0],  5.4126, 1e-4);
+      EXPECT_NEAR(inertia[0][2], 0, g_tolerance);
+      EXPECT_NEAR(inertia[2][0], 0, g_tolerance);
+      EXPECT_NEAR(inertia[1][2], 0, g_tolerance);
+      EXPECT_NEAR(inertia[2][1], 0, g_tolerance);
+    }
+
+    // For 0-2, apply torque and expect equivalent response
+    if (i <= 2)
+    {
+      for (int step = 0; step < 50; ++step)
+      {
+        link->SetTorque(math::Vector3(100, 0, 0));
+        world->Step(1);
+      }
+      if (_physicsEngine.compare("dart") == 0)
+      {
+        gzerr << "Dart fails this portion of the test (#1090)" << std::endl;
+      }
+      else
+      {
+        math::Vector3 vel = link->GetWorldAngularVel();
+        EXPECT_NEAR(vel.x,  0.0703, g_tolerance);
+        EXPECT_NEAR(vel.y, -0.0049, g_tolerance);
+        EXPECT_NEAR(vel.z,  0.0000, g_tolerance);
+      }
+    }
   }
 }
 
@@ -154,9 +406,21 @@ void PhysicsLinkTest::SetVelocity(const std::string &_physicsEngine)
 }
 
 /////////////////////////////////////////////////
+TEST_P(PhysicsLinkTest, GetWorldAngularMomentum)
+{
+  GetWorldAngularMomentum(GetParam());
+}
+
+/////////////////////////////////////////////////
 TEST_P(PhysicsLinkTest, GetWorldEnergy)
 {
   GetWorldEnergy(GetParam());
+}
+
+/////////////////////////////////////////////////
+TEST_P(PhysicsLinkTest, GetWorldInertia)
+{
+  GetWorldInertia(GetParam());
 }
 
 /////////////////////////////////////////////////